* origin/topic/matthias/bloom-filter:
  Use Bro-style platform-independent integer types.
  Change bloom filter's dependence on size_t.
  Remove debugging code.
  Update baseline with now correct FP tests.
  Add debugging code to find FP inconsistency.

Conflicts:
	src/3rdparty

+++ src/probabilistic/CounterVector.cc
@@ -153,7 +153,7 @@ CounterVector operator|(const CounterVector& x, const CounterVector& y)
 
 }
 
-size_t CounterVector::Hash() const
+uint64 CounterVector::Hash() const
 	{
 	return bits->Hash();
 	}
